+++ src/lib/parse.ts
@@ -1,6 +1,3 @@
-import { parseLS } from "./parse/LS";
-import { parsePrinceton } from "./parse/Princeton";
-
 export enum XUnits {
   Oe = "Oe",
   Am = "A/m",
@@ -53,18 +50,3 @@ export enum FileType {
 }
 
 export const FILE_TYPES = Object.values(FileType);
-
-export const parseFile = (
-  name: string,
-  source: string,
-  ty: FileType
-): ParseResult | null => {
-  switch (ty) {
-    case FileType.LS_DCD:
-    case FileType.LS_IRM:
-    case FileType.LS_HYST:
-      return parseLS(name, source, ty);
-    case FileType.PRINCETON:
-      return parsePrinceton(name, source, ty);
-  }
-};
